move utils and context over to connect-context, wrap with DHConnectProvider
